% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustMLE.R
\name{adjustMLE}
\alias{adjustMLE}
\title{Adjust MLE}
\usage{
adjustMLE(
  fit,
  method = c("SLOE", "ProbeFrontier"),
  check_MLE = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{fit}{glm object.}

\item{method}{a character. "SLOE" (default) or "ProbeFrontier".}

\item{check_MLE}{logical. If TRUE (default), check MLE exists.}

\item{quiet}{logical.}
}
\value{
modified glm object
}
\description{
Adjust MLE
}
