% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_eta.R
\name{estimate_eta_square}
\alias{estimate_eta_square}
\title{Estimate corrupted signal strength}
\usage{
estimate_eta_square(fit, X)
}
\arguments{
\item{fit}{glm object}

\item{X}{input data}
}
\value{
estimated eta squared
}
\description{
Estimate corrupted signal strength
}
