% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gamma.R
\name{estimate_gamma}
\alias{estimate_gamma}
\title{Estimate gamma by solving equation [4]}
\usage{
estimate_gamma(kappa_hat)
}
\arguments{
\item{kappa_hat}{double}
}
\value{
estimated gamma
}
\description{
Estimate gamma by solving equation [4]
}
