% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logistic}
\alias{logistic}
\title{Logistic function (a.k.a. sigmoid function)}
\usage{
logistic(q, location = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)
}
\description{
Logistic function (a.k.a. sigmoid function)
}
