% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logistic_derivative}
\alias{logistic_derivative}
\title{Derivative of logistic function}
\usage{
logistic_derivative(x, location = 0, scale = 1, log = FALSE)
}
\description{
Derivative of logistic function
}
