% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_kappa_hat.R
\name{search_kappa_hat}
\alias{search_kappa_hat}
\title{Search kappa_hat by Probing the MLE Frontier.
For details, see section 5.a.}
\usage{
search_kappa_hat(
  X,
  y,
  quiet = FALSE,
  B = 50L,
  kappa_step = 0.001,
  seed = NULL,
  solver = "lpsolve",
  n_cores = parallel::detectCores() - 1L
)
}
\arguments{
\item{X}{data.frame or matrix}

\item{y}{vector or matrix}

\item{quiet}{logical}

\item{B}{integer}

\item{kappa_step}{double}

\item{seed}{integer}

\item{solver}{character}

\item{n_cores}{integer}
}
\value{
kappa_hat
}
\description{
Search kappa_hat by Probing the MLE Frontier.
For details, see section 5.a.
}
