% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_the_system_of_equations.R
\name{solve_the_system_of_equations}
\alias{solve_the_system_of_equations}
\title{Solve the system of three non-linear equations.
Equation [5] from Sur and Candès (2018).
Equation (2.6) from Yadlowsky et al. (2021).}
\usage{
solve_the_system_of_equations(
  Sigma,
  kappa,
  initial_values = c(2, 1, 1),
  tol = 0.08
)
}
\arguments{
\item{Sigma}{function to calculate covariance matrix}

\item{kappa}{numeric. called "dimensionality" or "aspect ratio"}

\item{initial_values}{numeric vector of three elements}

\item{tol}{numeric. tolerance of integral}
}
\value{
solution contains x = [alpha, sigma_squared, lambda]
}
\description{
Solve the system of three non-linear equations.
Equation [5] from Sur and Candès (2018).
Equation (2.6) from Yadlowsky et al. (2021).
}
