% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_the_system_of_equations_with_eta.R
\name{solve_the_system_of_equations_with_eta}
\alias{solve_the_system_of_equations_with_eta}
\title{Solve the system of non-linear equations with gamma.
Equation (2.6) from Yadlowsky et al. (2021).
For details, see section 2.2.}
\usage{
solve_the_system_of_equations_with_eta(
  eta2_hat,
  kappa,
  initial_values = c(2, 1, 1),
  tol = 0.08
)
}
\arguments{
\item{kappa}{numeric. aspect ratio d/n}

\item{initial_values}{numeric vector of three elements}

\item{tol}{numeric. tolerance of integral}

\item{eta_hat}{numeric. estimated corrupted signal strength}
}
\value{
solution contains x = [alpha, sigma_squared, lambda]
}
\description{
Solve the system of non-linear equations with gamma.
Equation (2.6) from Yadlowsky et al. (2021).
For details, see section 2.2.
}
