% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_the_system_of_equations_with_gamma.R
\name{solve_the_system_of_equations_with_gamma}
\alias{solve_the_system_of_equations_with_gamma}
\title{Solve the system of non-linear equations with gamma.
Equation [5] as presented by Sur and Candès (2018).
For details, see section 4.b.}
\usage{
solve_the_system_of_equations_with_gamma(
  gamma_hat,
  kappa,
  initial_values = c(2, 1, 1),
  tol = 0.08
)
}
\arguments{
\item{gamma_hat}{numeric. estimated signal strength}

\item{kappa}{numeric. dimensionality p/n}

\item{initial_values}{numeric vector of three elements}

\item{tol}{numeric. tolerance of integral}
}
\value{
solution contains x = [alpha, sigma_squared, lambda]
}
\description{
Solve the system of non-linear equations with gamma.
Equation [5] as presented by Sur and Candès (2018).
For details, see section 4.b.
}
